Improve channel handling in deployCDK
Juju's handling of channels for bundles is strange. There is no way to
tell it to use a specific channel for a bundle without also forcing all
of the charms within the bundle to that same channel.

This refactors the `bundle_channel` param to be two separate params:

* `bundle_channel` Only influences the channel of the bundle itself
* `charms_channel` Forces the channel of all charms in the bundle

This is specifically needed for the Vault tests because the edge channel
of the Vault charm is not available.
(See [build #11](https://jenkins.canonical.com/k8s/job/validate-vault-v1.12.x/11/console))